Move WikidataOverviewStats to common, remove conditional display of total revisions
